// gravitee-apim-portal-webui/projects/portal-webclient-sdk/src/lib/model/ticketsResponse.ts
/**
 * Gravitee.io Portal Rest API
 * API dedicated to the devportal part of Gravitee
 *
 * The version of the OpenAPI document: 3.14.0-SNAPSHOT
 * Contact: contact@graviteesource.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Ticket } from './ticket';
import { Links } from './links';


export interface TicketsResponse { 
    /**
     * List of tickets.
     */
    data?: Array<Ticket>;
    /**
     * Map of Map of Object
     */
    metadata?: { [key: string]: { [key: string]: object; }; };
    links?: Links;
}